Move throttling to service.
